+++ src/main/java/cloneproject/Instagram/repository/post/PostRepository.java
@@ -1,4 +1,4 @@
-package cloneproject.Instagram.repository;
+package cloneproject.Instagram.repository.post;
 
 import cloneproject.Instagram.entity.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
